fixed asyn issues with proxy interface operations
Fixed async issues with proxy interface operations. All operations
should be functional and synchronized as neccessary. Reset and reload of
plugins only re-run plugins and will need to be revisited when the
plugin architecture is complete.

responses are posted to a redis channel as well as placed in redis with
the given key.