updated proxy pipline
Reworked the proxy processing pipline to properly handle request and
response manipulation via loaded modules.

Modules are loaded into the module registry. When a request event
occurs, the registry is notified. The registry then creates a request
processing pipline, then adds a callback function to make a client
connection. Similarly, when the client recieves a response event, the
client notifies the registry with a callback function. Once the response
has been processed by a generated response processing pipeline, the
callback is fired with the resulting response.