Updated session handling
Updated session handling to occure in the api call, rather than in the
module. When `start_session` is received the following occurs:

1. session document is retrieved from mongo
2. using the session doc, upstream host/port are set
3. session document is searched for profile id's, for every porfile id
found the corresponding module is started. Assuming it is not already
running. If ANY module is already running, then "busy" is returned.

Also changed some misc handling of respons messages